backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==4.2.18
django-cors-headers==4.6.0
django-import-export==4.3.3
django-rest==0.8.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tablib==3.7.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
numpy